add line cache

+++ aoc/incache.h
@@ -10,7 +10,7 @@ struct aoc_incache;
 struct aoc_incache *aoc_new_incache(char *pathname);
 void aoc_free_incache(struct aoc_incache *incache);
 size_t aoc_incache_size(struct aoc_incache *incache);
-int aoc_incache_get(struct aoc_incache *input_cache, int idx);
+int aoc_incache_get(struct aoc_incache *input_cache, size_t idx);
 
 #endif /* __AOC_INPUT_CACHE_H__ */
